Bump surya version
