+++ pyproject.toml
@@ -26,7 +26,7 @@ torch = "^2.5.1"
 tqdm = "^4.66.1"
 ftfy = "^6.1.1"
 rapidfuzz = "^3.8.1"
-surya-ocr = "~0.12.1"
+surya-ocr = "~0.13.0"
 regex = "^2024.4.28"
 pdftext = "~0.6.1"
 markdownify = "^0.13.1"
